+++ packages/editor/js/extensions/hooks/block-settings.js
@@ -4,6 +4,7 @@
  */
 import { select } from '@wordpress/data';
 import type { MixedElement } from 'react';
+import { doAction } from '@wordpress/hooks';
 import { memo, useEffect } from '@wordpress/element';
 import { SlotFillProvider, Slot } from '@wordpress/components';
 
@@ -41,7 +42,6 @@ import {
 	BlockIcon,
 	propsAreEqual,
 } from '../components';
-import { bootstrapCanvasEditor } from '../../canvas-editor';
 
 const useSharedBlockSideEffect = (): void => {
 	const {
@@ -103,8 +103,22 @@ export default function withBlockSettings(
 	args: extraArguments
 ): Object {
 	const { getBlockExtensionBy } = select(STORE_NAME) || {};
+	const { getExtension } = select('blockera/extensions/config');
 
 	const blockExtension = getBlockExtensionBy('targetBlock', name);
+
+	const { registerExtensions = null } = blockExtension || {};
+
+	// On registering block type, we're firing bootstrapper scripts and add experimental extensions support.
+	if ('function' === typeof registerExtensions) {
+		registerExtensions(settings.name);
+	} else if (!getExtension(settings.name)) {
+		registerBlockExtensionsSupports(settings.name);
+		registerInnerBlockExtensionsSupports(
+			settings.name,
+			blockExtension?.blockeraInnerBlocks || {}
+		);
+	}
 
 	if (blockExtension && isBlockTypeExtension(blockExtension)) {
 		return mergeBlockSettings(settings, blockExtension, args);
@@ -143,7 +157,6 @@ function mergeBlockSettings(
 		getSharedBlockAttributes = () => ({}),
 		getBlockTypeAttributes = () => ({}),
 	} = select('blockera/extensions') || {};
-	const { getExtension } = select('blockera/extensions/config');
 
 	const isAvailableBlock = () =>
 		!unsupportedBlocks.includes(settings.name) &&
@@ -192,19 +205,6 @@ function mergeBlockSettings(
 		? getSharedBlockAttributes()
 		: blockeraOverrideBlockTypeAttributes;
 
-	const { registerExtensions = null } = additional;
-
-	// On registering block type, we're firing bootstrapper scripts and add experimental extensions support.
-	if ('function' === typeof registerExtensions) {
-		registerExtensions(settings.name);
-	} else if (!getExtension(settings.name)) {
-		registerBlockExtensionsSupports(settings.name);
-		registerInnerBlockExtensionsSupports(
-			settings.name,
-			additional?.blockeraInnerBlocks || {}
-		);
-	}
-
 	const Edit = memo((props: Object): MixedElement => {
 		if (isFunction(additional?.edit) && isAvailableBlock()) {
 			const baseContextValue = {
@@ -216,7 +216,10 @@ function mergeBlockSettings(
 
 			// Bootstrap canvas editor UI on WordPress site editor.
 			if (!isLoadedPostEditor() && isLoadedSiteEditor()) {
-				bootstrapCanvasEditor('site');
+				/**
+				 * Calls the callback functions that have been added to an action hook.
+				 */
+				doAction('blockera.mergeBlockSettings.Edit.component');
 			}
 
 			return (