+++ packages/editor/js/extensions/libs/effects/test/backdrop-filters.e2e.cy.js
@@ -86,7 +86,7 @@ describe('Backdrop Filters → Functionality', () => {
 
 		redirectToFrontPage();
 
-		cy.get('style#blockera-inline-css-inline-css')
+		cy.get('style#blockera-inline-css')
 			.invoke('text')
 			.should(
 				'include',