+++ packages/editor/js/extensions/libs/typography/test/font-size.e2e.cy.js
@@ -83,7 +83,7 @@ describe('Font Size → Functionality', () => {
 
 		redirectToFrontPage();
 
-		cy.get('style#blockera-inline-css-inline-css')
+		cy.get('style#blockera-inline-css')
 			.invoke('text')
 			.should(
 				'include',