+++ packages/editor/js/hooks/use-attributes/index.js
@@ -32,6 +32,7 @@ export const useAttributes = (
 	setAttributes: (attributes: Object) => void,
 	{
 		blockId,
+		clientId,
 		className,
 		innerBlocks,
 		currentBlock,
@@ -49,6 +50,7 @@ export const useAttributes = (
 		getActiveBlockVariation,
 	}: {
 		blockId: string,
+		clientId: string,
 		className: string,
 		innerBlocks: Object,
 		currentState: TStates,
@@ -98,8 +100,7 @@ export const useAttributes = (
 	): void => {
 		const { ref, effectiveItems = {} } = options;
 		const { getSelectedBlock } = select('core/block-editor');
-		const { attributes = getAttributes(), clientId } =
-			getSelectedBlock() || {};
+		const { attributes = getAttributes() } = getSelectedBlock() || {};
 
 		// attributes => immutable - mean just read-only!
 		// _attributes => mutable - mean readable and writable constant!
@@ -118,7 +119,10 @@ export const useAttributes = (
 
 		// Sets "blockeraPropsId" if it is empty.
 		if (!_attributes?.blockeraPropsId) {
-			_attributes = getAttributesWithIds(_attributes, 'blockeraPropsId');
+			_attributes = {
+				..._attributes,
+				blockeraPropsId: clientId,
+			};
 		}
 
 		const indexOfBlockeraSelector =