+++ packages/editor/CHANGELOG.md
@@ -15,6 +15,7 @@
 - Wrong style generation for blocks in gap feature [[🔗 Bug](https://community.blockera.ai/bugs-mdhyb8nc/post/wrong-style-generation-for-blocks-in-gap-feature-nkRSbYvjaK226Rh)]
 - Refactored BlockBase component to fix compatibility with WordPress problems. [[🔗 Bug](https://community.blockera.ai/bugs-mdhyb8nc/post/group-block-variation-control-bug-with-updating-layout-section---flex-MSd1lhFdxAGDd2c)]
 - Fixed an issue where advanced values of the Flex Layout (e.g., `space-between`, `space-around`) were not working properly. [[🔗 Link](https://community.blockera.ai/bugs-mdhyb8nc/post/advanced-values-of-flex-layout-not-working-properly-wTCfgyDW4w1EoVK)]
+- Fixed a style issue when the selected block is unsupported. [[🔗 Bug](https://community.blockera.ai/bugs-mdhyb8nc/post/style-issue-if-current-selected-block-is-a-not-supported-block-9VjcCa2CSA7FPpl)]
 - Fixed an issue that caused duplicate block cards to be displayed. [[🔗 Bug](https://community.blockera.ai/bugs-mdhyb8nc/post/extra-block-card-in-the-settings-tab-0C3lR6niX6ygYir)]
 
 ### Improvements
@@ -22,10 +23,13 @@
 - Design of post preview link in header improved.
 - Enhanced the getNormalizedSelector API for better performance and functionality.
 - Refactored server-side style engine with cleaner code and improved support for & ampersand in selector creation.
-- Added spacer for processed selector in the client-side style engine to improved css selector normalizer. 
+- Added spacer for processed selector in the client-side style engine to improved css selector normalizer.
+- Added a reset method in Layout style definition.
 
 ### Automated Tests:
 - Completed E2E tests to verify the functionality of all advanced states in Flex Layout.
 - Conducted WP data compatibility test for the Spacer Block to ensure correct functionality.
 - Added new test suites to ensure the robustness of the getNormalizedSelector API.
 - Updated tests to detect real hover states more accurately.
+- Updated Blockera inline style tag ID to ensure all assertions pass successfully.
+- Added new automated tests to verify unique class names for blocks in Blockera.